major rev of mount_smb
mount_smb is now robust in the face of network outages.  It now
properly waits for autossh to reestablish the ssh connection to the
sshhost.

Also added "debug" mode.  If $debug is non-null, then informative,
date-based messages are printed, so let you know what mount_smb is
doing.

New feature: if defined, a function mount_smb_timer is now called
every minute with a single argument, the number of minutes since first
mount, so knocking can be done, for example.