mount_smb: --debug renamed to --verbose; tweaks
--debug was a bad name for the argument.  It's now --verbose.

Make sure password is asked for again if the initial mount fails.
That was a bug in the last commit.

Add some potential countermeasures for the autossh seeming to fail, by
never creating the pidfile.  Wait longer for it before giving up.

The rest of the changes are cosmetic and documentation.