bash_sshagent: get rid of `info2' nonsense
Saving SSH_AUTH_SOCK for use on machines were it there was no agent
running, and assuming it was useful was a very wrong assumption.  The
value of SSH_AUTH_SOCK is only good for the duration of the ssh
command that set it.  Once that command exits, the SSH_AUTH_SOCK value
is no longer valid.  It does no good to save this value and use it
later.  Remove all the code that did this.