bash_sshagent: __my_ssh_ensure_default_identities change
The arrays __my_ssh_default_identity and __my_ssh_default_identity_md5
contain a matched list of SSH identities and MD5 fingerprints.  They
are used to ensure that the SSH keys are added to the running agent.
The marker "done", in __my_ssh_default_identity, when a corresponding
match is made in __my_ssh_default_identity_md5, causes the loop through
the arrays to terminate.  This is useful when using FastX via a tunnel,
using a forwarded agent key, the file of which is not accessible locally
(but the key is in ~/.ssh/authorized_keys).